store/requirements.txt
boto3==1.26.98
botocore==1.29.98
dj-config-url==0.1.1
Django==4.1.7
django-extensions==3.2.1
django-storages==1.13.2
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Faker==18.3.1
Pillow==9.4.0
psycopg2-binary==2.9.5
s3transfer==0.6.0
sqlparse==0.4.3
urllib3==1.26.15
